Removed EventRole from user model and added it to enrollment model
