+++ lms/models/__init__.py
@@ -1,4 +1,4 @@
 from .event_models import Event, EventInstance
-from .user_models import User, EventRole
-from .enrollment_models import UserEnrollment
+from .user_models import User
+from .enrollment_models import EventRole, UserEnrollment
 from .feedback_models import EventInstanceFeedback
