+++ lms/models/enrollment_models.py
@@ -1,6 +1,13 @@
 from django.db import models
-from . import User, EventRole, EventInstance
+from . import User, EventInstance
 
+
+class EventRole(models.IntegerChoices):
+    PARTICIPANT = 1
+    FACILITATOR = 2
+    EVENT_ADMIN = 3
+    COORDINATOR = 4
+    LEAD = 5
 
 class UserEnrollment(models.Model):
     ''' Store administrative information about participant'''
